+++ realtime/src/config.rs
@@ -1,5 +1,5 @@
 use std::ops::RangeInclusive;
-pub use xsynth_core::channel::ChannelInitOptions;
+pub use xsynth_core::{channel::ChannelInitOptions, channel_group::ThreadCount};
 
 /// Options for initializing a new RealtimeSynth.
 pub struct XSynthRealtimeConfig {
@@ -25,17 +25,12 @@ pub struct XSynthRealtimeConfig {
     /// Default: `[9]`
     pub drums_channels: Vec<u32>,
 
-    /// Controls the use a threadpool to render individual keys' voices.
-    /// When a value is set, the specified number of threads will be used
-    /// for that operation, while `None` means no per-key concurrency.
-    /// If the value is set to `Some(0)` then the number of threads will
-    /// be determined automatically by `rayon`.
+    /// Controls the multithreading used for rendering per-voice audio for all
+    /// the voices stored in a key for a channel. See the `ThreadCount` documentation
+    /// for the available options.
     ///
-    /// Regardless, each MIDI channel uses its own thread. This setting
-    /// adds more fine-grained threading per key rather than per channel.
-    ///
-    /// Default: `None`
-    pub threadpool: Option<usize>,
+    /// Default: `ThreadCount::None`
+    pub multithreading: ThreadCount,
 
     /// A range of velocities that will not be played.
     ///
@@ -50,7 +45,7 @@ impl Default for XSynthRealtimeConfig {
             render_window_ms: 10.0,
             channel_count: 16,
             drums_channels: vec![9],
-            threadpool: None,
+            multithreading: ThreadCount::None,
             ignore_range: 0..=0,
         }
     }
